+++ src/test/java/apitestscenarios/CreateUserWithPostApiAuthRegisterTest.java
@@ -1,4 +1,4 @@
-package apiTestScenarios;
+package apitestscenarios;
 
 import com.github.javafaker.Faker;
 import io.qameta.allure.Step;
@@ -7,13 +7,13 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import serialsClasses.ApiEndPoints;
-import serialsClasses.UserSerials;
+import serialsclasses.ApiEndPoints;
+import serialsclasses.UserSerials;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class ChangeUserWithPatchApiAuthUserTest {
+public class CreateUserWithPostApiAuthRegisterTest {
     UserSerials user;
     ApiEndPoints api = new ApiEndPoints();
     String email;
@@ -29,33 +29,16 @@ public class ChangeUserWithPatchApiAuthUserTest {
         password = faker.internet().password();
         name = faker.name().firstName();
         RestAssured.baseURI = "https://stellarburgers.nomoreparties.site";
-        response = sendRequest(email, password, name);
-        token = getToken(response).replaceFirst("Bearer ", "");
     }
 
-    @Step("Send POST request")
+    @Step ("Send POST request")
     public Response sendRequest(String email, String password, String name){
         user = new UserSerials(email, password, name);
-
         response = given()
                 .header("Content-type", "application/json")
                 .body(user)
                 .when()
                 .post(api.getCreateUser());
-        return response;
-    }
-
-    @Step("Send PATCH request")
-    public Response sendPatchRequest(String token, String email, String password, String name){
-        user = new UserSerials(email, password, name);
-
-        response = given()
-                .header("Content-type", "application/json")
-                .auth().oauth2(token)
-                .and()
-                .body(user)
-                .when()
-                .patch(api.getChangeUser());
         return response;
     }
 
@@ -68,32 +51,51 @@ public class ChangeUserWithPatchApiAuthUserTest {
     }
 
     @Step ("Get Token")
-    public String getToken(Response response) {
+    public String getToken(Response response){
         return response.then().extract().body().path("accessToken");
     }
 
     @Test
-    public void canChangeUserWithAuthorisation(){
-        response = sendPatchRequest(token,email + "ru", password + "123", name + "s");
+    public void canCreateUser(){
+        response = sendRequest(email, password, name);
+
         checkStatus (response, true, "200 OK");
     }
 
     @Test
-    public void cannotChangeUserWithoutAuthorisation(){
-        response = sendPatchRequest("", email + "ru", password + "123", name + "s");
-        checkStatus (response, false,"401 Unauthorized");
+    public void cannotCreateTwoIdenticalUsers(){
+        response = sendRequest(email, password, name);
+        Response response1 = sendRequest(email, password, name);
+
+        checkStatus (response1, false, "403 Forbidden");
+
     }
 
     @Test
-    public void cannotChangeUserWithExistEmail(){
-        sendRequest(name + email, password, name);
-        response = sendPatchRequest(token,name + email, password, name);
+    public void cannotCreateUserWithoutEmail(){
+        response = sendRequest("", password, name);
+
+        checkStatus (response, false, "403 Forbidden");
+    }
+
+    @Test
+    public void cannotCreateUserWithoutPassword(){
+        response = sendRequest(email, "", name);
+
+        checkStatus (response, false,"403 Forbidden");
+    }
+
+    @Test
+    public void cannotCreateUserWithoutName(){
+        response = sendRequest(email, password, "");
+
         checkStatus (response, false,"403 Forbidden");
     }
 
     @After
     public void clearAll(){
         try {
+            token = getToken(response).replaceFirst("Bearer ", "");
             given()
                     .auth().oauth2(token)
                     .delete(api.getChangeUser());
@@ -101,4 +103,5 @@ public class ChangeUserWithPatchApiAuthUserTest {
             token = "";
         }
     }
+
 }
