+++ src/test/java/apitestscenarios/GetUsersOrdersWithGetApiOrdersTest.java
@@ -1,4 +1,4 @@
-package apiTestScenarios;
+package apitestscenarios;
 
 import com.github.javafaker.Faker;
 import io.qameta.allure.Step;
@@ -7,19 +7,23 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import serialsClasses.ApiEndPoints;
-import serialsClasses.UserSerials;
+import serialsclasses.ApiEndPoints;
+import serialsclasses.UserSerials;
+import java.util.ArrayList;
+import java.util.List;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class LoginUserWithPostApiAuthLoginTest {
+public class GetUsersOrdersWithGetApiOrdersTest {
+
     UserSerials user;
     ApiEndPoints api = new ApiEndPoints();
     String email;
     String password;
     String name;
+    List<String> ingredients = new ArrayList<>();
     Response response;
     String token = "";
 
@@ -29,13 +33,18 @@ public class LoginUserWithPostApiAuthLoginTest {
         email = faker.internet().emailAddress();
         password = faker.internet().password();
         name = faker.name().firstName();
+        ingredients.add("61c0c5a71d1f82001bdaaa6d");
+        ingredients.add("61c0c5a71d1f82001bdaaa6f");
         RestAssured.baseURI = "https://stellarburgers.nomoreparties.site";
         response = sendRequest(email, password, name);
+        token = getToken(response).replaceFirst("Bearer ", "");
+        sendPostOrderRequest(token, ingredients);
     }
 
-    @Step("Send POST request /api/auth/register")
+    @Step("Send POST request")
     public Response sendRequest(String email, String password, String name){
         user = new UserSerials(email, password, name);
+
         response = given()
                 .header("Content-type", "application/json")
                 .body(user)
@@ -44,14 +53,26 @@ public class LoginUserWithPostApiAuthLoginTest {
         return response;
     }
 
-    @Step("Send POST request /api/auth/login")
-    public Response sendLoginRequest(String email, String password){
-        user = new UserSerials(email, password);
+    @Step("Send POST order request")
+    public Response sendPostOrderRequest(String token, List<String> ingredients){
+        user = new UserSerials(ingredients);
+
         response = given()
                 .header("Content-type", "application/json")
+                .auth().oauth2(token)
+                .and()
                 .body(user)
                 .when()
-                .post(api.getLoginUser());
+                .post(api.getCreateOrder());
+        return response;
+    }
+
+    @Step("Send GET order request")
+    public Response sendGetOrderRequest(String token){
+        response = given()
+                .header("Content-type", "application/json")
+                .auth().oauth2(token)
+                .get(api.getCreateOrder());
         return response;
     }
 
@@ -64,36 +85,26 @@ public class LoginUserWithPostApiAuthLoginTest {
     }
 
     @Step ("Get Token")
-    public String getToken(Response response){
+    public String getToken(Response response) {
         return response.then().extract().body().path("accessToken");
     }
 
     @Test
-    public void canLoginExistUser(){
-        response = sendLoginRequest(email, password);
-
+    public void canGetOrdersWithAuthUser(){
+        response = sendGetOrderRequest(token);
         checkStatus (response, true, "200 OK");
     }
 
     @Test
-    public void cannotLoginUserWithWrongEmail(){
-        response = sendLoginRequest(email + "ru", password);
-
-        checkStatus (response, false, "401 Unauthorized");
-    }
-
-    @Test
-    public void cannotLoginUserWithWrongPassword(){
-        response = sendLoginRequest(email, password + "123");
-
+    public void cannotGetOrdersWithoutAuthUser(){
+        response = sendGetOrderRequest("");
         checkStatus (response, false, "401 Unauthorized");
     }
 
     @After
     public void clearAll(){
         try {
-            token = getToken(response).replaceFirst("Bearer ", "");
-            response = given()
+            given()
                     .auth().oauth2(token)
                     .delete(api.getChangeUser());
         } catch (Exception exception) {
