+++ src/test/java/serialsclasses/UserSerials.java
@@ -1,4 +1,4 @@
-package serialsClasses;
+package serialsclasses;
 
 
 import java.util.List;
